added deque to provide proxies
changed proxy storage to deque storing them in order of descending time of timeout. From this storage proxies are given from the left and put in a sorted way. Main storage remains a dict of all proxies that stores all proxies provided to proxy pool

This required removal of multiple thread proxies